moved doc api to helpers and made updates to docinfo component
