+++ app/components/docinfo.js
@@ -9,23 +9,34 @@ class App extends React.Component {
 
 	};
 
-	const api_key = '4e6764b7fa2910445e23428d12cdc643';
-	// any functions that we want to add.  i.e changing the state
-
-	function doctorSearch() {
-		searchVal = $("This is tbd once the frontend is starting to get worked on").val().trim();
-		var resource_url = 'https://api.betterdoctor.com/2016-03-01/doctors?location=37.773,-122.413,100&skip=2&limit=10&user_key=' + api_key;
-		baseAjax = $.ajax({
-			url: resource_url,
-			method: 'GET',
-			dataType: 'jsonp'
-		}).done(function(response) )
-
-
-
 	}
+	render() {
+    return (
+
+      <div className="container">
+
+        <div className="row">
+
+          <div className="col-lg-12">
+
+            <div className="panel panel-default">
+              <div className="panel-heading">
+                <h3 className="panel-title">Disease: {this.props.route.disease}</h3>
+              </div>
+              <div className="panel-body">
+                The disease is {this.props.route.disease}
+              </div>
+              <Forum disease={this.props.route.disease} />
+              <Scrape />
+            </div>
+
+          </div>
+
+        </div>
+      </div>
+    );
+  }
 
 };
 
-ReactDOM.render(
-  <App />, document.getElementById("app"));
+export default DocInfo
